get readwriterec building by moving some more functions around
