+++ services/src/ispf.c
@@ -69,71 +69,3 @@ int ispf_disk_stats_to_ispf_stats(const char* userdata, int userdata_len, struct
     
   return 0;
 }
-
-static char* PTR32 ispf_rname(const char* ds, const char* mem)
-{
-  unsigned int rname_len = strlen(ds) + strlen(mem);
-
-  if (rname_len > 44+8) {
-    fprintf(stderr, "Invalid dataset or member name passed to ENQ/DEQ %s(%s)\n", ds, mem);
-    return NULL;
-  }
-
-  char* PTR32 rname;
-  rname = MALLOC31(52+1);
-  if (!rname) {
-    fprintf(stderr, "Unable to obtain storage for ENQ/DEQ\n");
-    return NULL;
-  }
-  sprintf(rname, "%-44s%-8s", ds, mem);
-
-  return rname;
-}
-
-static char* PTR32 ispf_qname(const char* qn)
-{
-  unsigned int qname_len = strlen(qn);
-
-  if (qname_len > 8) {
-    fprintf(stderr, "Invalid queue name passed to ENQ/DEQ %s\n", qn);
-    return NULL;
-  }
-
-  char* PTR32 qname;
-  qname = MALLOC31(8+1);
-  if (!qname) {
-    fprintf(stderr, "Unable to obtain storage for ENQ/DEQ\n");
-    return NULL;
-  }
-  sprintf(qname, "%-8s", qn);
-
-  return qname;
-}
-
-int ispf_enq_dataset_member(const char* ds, const char* wmem) 
-{
-  char* PTR32 rname = ispf_rname(ds, wmem);
-  char* PTR32 qname = ispf_qname("SPFEDIT");
-
-  if (!rname || !qname) {
-    return 4;
-  }
-  int rc = SYEXENQ(qname, rname, strlen(rname));
-  free(rname);
-  free(qname);
-  return rc;
-}
-
-int ispf_deq_dataset_member(const char* ds, const char* wmem) 
-{
-  char* PTR32 rname = ispf_rname(ds, wmem);
-  char* PTR32 qname = ispf_qname("SPFEDIT");
-
-  if (!rname || !qname) {
-    return 4;
-  }
-  int rc = SYEXDEQ(qname, rname, strlen(rname));
-  free(rname);
-  free(qname);
-  return rc;
-}
